+++ packages/web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @render-harness/web
 
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @render-harness/registry@0.5.1
+
 ## 0.5.0
 
 ### Minor Changes